Complete refactor of how expectations are handled
Expects are now run at the call site instead of delaying until dsl
is complete.
This was necessary in order to create certain methods like reified inline calls for
expectException and will most likely be needed for future calls as well.
This has the added benefit of failing immediately instead of delaying until 'build' time